requirements.txt
numpy==1.20.1
Pillow==8.1.1
requests==2.25.1
pytest==6.2.2
